Always copy files as writable (#15386)
Use `chmod` to ensure that any copied files are writable. I tried using
`copyfile` which avoids copying permissions altogether by we have some
files (e.g. `sdl2-config`) which are executable so we need to preserve
at least some permissions.

This can be importantant in `install_system_headers` which copies files
from the emscripten directory (which may be read only) in the cache
directory. In this case we don't want the resulting files in the cache
to also be read only.

Fixes: #15374