Use `--export-dynamic` rather than `--export-all` when linking with MAIN_MODULE=1 (#15413)
This means hidden symbols no longer get exported by default.  The
only way to export a hidden symbol when building with MAIN_MODULE,
SIDE_MODULE or LINKABLE is to add it to the EXPORT_FUNCTIONS list.

Also, update the `test_dlfcn_self` test which check the number of data
exports to be precise.  This means it will need updating anytime libc or
some other core library adds or removes a data exports, but that seems
like something we should be tracking here anyway.