Add scripts for copying changes out back into upstream llvm and musl (#15430)
We are now maintaining our llvm and musl patches in external
repositories so that it is easier to update from upstream
and rebase onto new releases.

I already created scripts for copying changes in one direction
(i.e. into emscripten).  This change adds push_llvm_changes.py
and push_musl_changes.py which can be used to push changes
in the other direction.