Add some new musl math files to libc-rt library
We have had reports of missing symbols in LTO builds after
the musl upgrade (#13006):
https://groups.google.com/g/emscripten-discuss/c/g_vkRRSCPUI/m/aT6fnUAtCQAJ?utm_medium=email&utm_source=footer

Also `lto2.test_float_builtinsTest` started failing on the
emscripten-releases FYI waterfall.

This change moves some of the new math-related source files
into the libc-mt library which is excluded from LTO (see comments
in in system_libs.py for background on this).

Fixes: #15506