Free thread data in native code. NFC
Thread data is now free'd in native code and no longer requires a
postMessage back to the main thread.  We free the thread data either
when the thread exits (in the case it is detached) or when it is joined
(in the case is not detached).

In either case we can re-cycle the worker (return it to the worker pool)
right away since the thread is no longer running.

Since `freeThreadData` is no longer called during `returnWorkerToPool`
it means that pthreads that are terminated asyncronously (with
prejudice) will not leak their thread data.  However terminating a
thread in this way (using worker.terminate()) is already racey and
dangerous since we don't know what state the diing thread is in, or
exactly when it is safe to free its data.