Map printErr to console.warn rather then console.error
This matches the default mapping of the `err` JS helper function.

This change allows us to correctly run test_emscripten_console_log.
This test was previously not running correctly but this was going
unnotived due to mismatch in mapping.

Alternatively, we could go the other direction and change `err` to map to
console.error, but I think it makes sense the defulat `err` and the
default `printErr` map to the same thing.