Avoid overriding printErr in example shell html
The default version of printErr already writes to the console so the
override is not actually doing anything useful.

However because the override is writing to console.error rather than
console.warn which is the default in shell.js this override can actually
cause some confusion.  Specifcally the test_emscripten_console_log test
expects `err` to be going to a certain place and this override was
meaning the place it was going was different when building with the html
shell vs the default.

As an alternative we should leave these overrides in as an example but
use console.warn instead.  I'm inclinded not to do that though since is
add complexity to these shell files.