Simplify locking strategy
The previous solution ensured that the worker always either held the lock or was
waiting on the condition variable to prevent invokers from overwriting each
other's submitted work. However, invokers waiting for the state to return to
`Waiting` before submitting their work already protects against this failure
mode. As a result, we can simplify the locking strategy to only hold locks when
state needs to be accessed or modified. We won't miss important wake up
notifications because we are always atomically checking the state before
waiting; if we miss a notification, we won't end up waiting at all.

As a result of this change:

 - We no longer need to wait for the worker thread to start up in the constructor.
 - We no longer need `workerLock`

Also make some unrelated fixes:

 - Remove redundant check for `state != Waiting` before waiting in `invoke`.
 - Better handling of overflow of `workCount`