Build a non-debug version of libc
We have a fair amount of assert usage in our libc code (in parituclar
in our custom pthread code).  This change is a codesize win for release
builds that can avoid including that.

I would like also use NDEBUG to avoid calling out to
emscripten_thread_set_name/status in release builds in the future.

See #15628