Clarify setThreadStatus by removing first argument. NFC
This functions only ever operates on the current and completely removing
the first argument helps to clarify this.

The native versions already don't take a thread as an argument in assume
they are operating on the current thread.

This function was operating with mix of atomic and non-atomic
operattions which made it unsafe except for running against the current
thread block.