Use Object.assign where available. NFC
The reason this doesn't show up in any code size tests is (I think)
because the of the two types of code size tests we have one only
measures with MINIMAL_RUNTIME (tests/code_size) and the other only looks
at the size of the wasm file (tests/metadce).  We might want to consider
increasing coverage so that this kind of win is picked up the tests?