[WASMFS] Fuzzer setup (#15689)
Relevant Issue: #15041 

Objective: Setup simple fuzzer to identify unexpected behaviour caused by data races and other thread-related issues in WasmFS.

A test should check for tearing in file reads and writes. If thread A is performing a write and then thread B performs a read, the read should not interrupt the write. Reads and writes should ensure data consistency (i.e. it should not be possible to read a file in an intermediate state and observe only partial data that has been written).