Mode atomics.wait path for `emscripten_futex_wait` to native code. NFC
Add a new WebAssembly global to the native side that signals if
the current thread can use `atomics.wait` and use this to implement
the `atomics.wait` version of `emscripten_futex_wait`.

The non-blocking/busy-looping path is still implemented in JS
for now.  It could potentially be moved as a followup.