Parse settings on the command line passed in as hex
I ran into this issue when working on #15763.  We have code interally
that converts `=-1` to `=0x7FFFFFFF` for the MIN_XX_VERSION settings.
However, this was then being interpreted interally as a string.  When
read from JS there is no differenence but from python code the value of
this setting would appear as the string "0x7FFFFFFF" and not the number
0x7FFFFFFF.