Ensure objects within library files are sorted
We saw an odd issue recently where clearing the cache and rebuilding
generated different result.  Looking that the library files I notices
that orders of the object files with each library did not match.

This change means the order will be fixed regardless of ordering of the
files in the directories of the host machine.

See #15762