Use polyfil rather for Object.assign
This gives a small code size saving and it makes our codebase
more readable.

The reason for this change is that I ran into an issue where I was
trying to use `objAssign` in more places and it became apparent
that MINIMAL_RUNTIME didn't define the `objAssign` helper.

Create a new directory for out polyfills, of which we now have
two.