Modernize internal JS code and run it through eslint. NFC
This change does not effect the JS that we output, only the the internal
JS tooling.

This is a followup to #15836 which did this cleanup just for a single
file.

The long list of exclusions in `.eslintrc.yml` is because these files
are part of the JS library code which we ship and we a lot of that
cannot be run though the lint tool because it uses our speciall
pre-processing techniques.