Avoid parsing settings.js files in src/compiler.js. (#15911)
We always pass the settings via a json file which is created by the
python code.   Additionally parsing the settings as JS here serves no
purpose that I can tell.

It might have been useful at some point in the past for debugging, but
I've certainly never used this feature.  It should be perfectly possible
to create a json file on demand in one needs one.

The motivation for doing this is that (a) it simplifies the code, (b)
it allows us to switch settings.js to python and (c) it moves us closer
to modernizing `src/compiler.js` by avoiding depending quite so much
on the legacy behaviour of `eval`.