[wasmfs] Make standalone's open syscall weak, to not conflict with wasmfs (#15992)
Both libwasmfs and libstandalonewasm implement the open syscall, as well as
others (only the first is in this PR). I think we should make them weak in
standalone wasm, that is, if wasmfs is present we always want to run that code.
The wasmfs syscall will then call lower-level things which we will need to implement
in standalone wasm (like we do in JS).