[WasmFS] Flush streams on exit to fix EXIT_RUNTIME behavior (#16029)
Without this, printf("hello world") without a newline at the end will print
nothing. In EXIT_RUNTIME mode the streams should be flushed.

The old FS calls fflush from JS. This adds a destructor to WasmFS which
flushes the standard streams. (This does not yet support the JS ability to
warn on unflushed output without EXIT_RUNTIME mode; we may want
to find a way to support that later.)

Also remove some trivial and unnecessary duplicated code by adding a
new WritingStdFile class, which I noticed while editing that file, and other
tiny cleanups there.

Add a flush() method to Files, which we may iterate on later, but for now
lets us flush the standard streams.