[WasmFS] Partial support for standalone mode with WasmFS (#16059)
Implement stubs for standalone things we do not expect to handle like
preloaded files. Handle stdout/stderr using interposition, importing the
VM's fd_write under a different name.

This is not complete (and so #16008 will not be closed) because
to actually run in a wasm VM we need to implement more syscalls in WasmFS.
Right now those end up as imports that the VM cannot handle. However, in
an optimized build some tests already pass (as those imports are removed),
so this PR is able to add some testing.

Also switch wasmfs to use -O2, which is consistent with other modes
(see the ones nearby in the diff), and which is necessary as just mentioned
for the testing added.