[WasmFS] Begin JS API integration (#16068)
This adds a few initial JS API functions. These will be enough to enable a test
in 2-3 PRs, and so no new test is added here.

The policy for this API is the same as the old FS: By default WasmFS includes
the JS API parts that it itself needs, and nothing more. If the user provides
FORCE_FILESYSTEM then the entire JS API is included. (As WasmFS uses
less JS, this flag may end up used more in practice, I'm not sure. But I thought
it best not to add a new flag here and keep the existing convention.)

Also rename the existing read_file JS API support function for consistency,
and move it to the new location. No changes to the function's contents.