Fix freeaddrinfo(), which did not match allocation in JS (#16085)
It is risky to allocate in JS and free in C, when given matching allocate/free
pairs like getaddrinfo/freeaddrinfo, and it looks like we had things wrong
here. That is, the existing musl code did not match how we use that
structure. The worst part is that musl appears to assume the struct is
identical/can alias some other struct, and that led to memory corruption.

Not sure if we can test this reliably as the issue is corruption.

Fixes #16081