[WasmFS] Initial symlink support (#16074)
This only implements enough of them to get the most basic test passing. It
defines the Symlink class, the symlink and readlink syscalls, and a JS API
for symlink. No following of symlinks or reading from them etc. is done yet.

This refactors test_unistd_links to be parameterized and then enables it
for wasmfs. While doing so I realized that the old FS appears to be wrong:
it normalizes paths in an odd way, that I don't see in the spec nor when I
run on linux. The test now expects different results based on which FS it
is. The WasmFS results match linux.

See #15948