Avoid reverse dependencies in JS library code handling time. NFC
Specifically this change removes `_get_daylight`, `_get_daylight`
and `_get_tzname` which are native symbols that were used by
`tzset` which has a lot of different call sites.

This made reverse dependencies on these symbols fairly common
and easy to miss.  See #15958 for a recent example.

This change removes these functions completely in favor or just
passing these addresses in as arguments.

Fixes: #15958