Deprecate `--bind` in favor of just `-lembind`. NFC (#16087)
All that the `--bind` flag does is enable the embind library (native
and JS portions), so I think its more clear if we just use the normal
`-l` method of including it.

Historically `--bind` has sounded to some (including myself) like a
verb rather than the name of a library leading to folks to believe that
the linker is actually doing the binding, whereas all the binding is
done either at compile time or at load time.  No "binding" happens
during linking.