[SjLj] Enable Wasm SjLj with -sSUPPORT_LONGJMP=wasm (#16111)
Because currently `DISABLE_EXCEPTION_THROWING` is 0 by default, which
enables Emscripten EH exception throwing, to use Wasm SjLj, you need to
specify both options, which is inconvenient and unintuitive:
```
emcc -sSUPPORT_LONGJMP=wasm -sDISABLE_EXCEPTION_THROWING=1 test.c
```

So far we've errored out when `SUPPORT_LONGJMP=wasm` was given with
either `DISABLE_EXCEPTION_CATCHING=0` or `DISABLE_EXCEPTION_THROWING=0`.
But unlike `DISABLE_EXCEPTION_CATCHING`, which is 1 by default and you
have to explicitly disable it to make it 0, `DISABLE_EXCEPTION_THROWING`
is 0 by default so you always have to turn it off when we use Wasm SjLj,
which is very inconvenient. This PR automatically sets
`DISABLE_EXCEPTION_THROWING` to 1 when `SUPPORT_LONGJMP` is `wasm`.