Remove unnecessary dependency on `stderr` debug builds. NFC
In debug builds this stub version was bringing in other dependencies via
`getpid` -> `__syscall_getpid` -> `stderr`.  Due to the way object file
dependecies get resolved this means that any dependency on any function
in `library_pthread_stub.c` was bringing in `stderr` which in turn adds
`__stdio_write/read/lseek` to the wasm table (which is not amenable to
DCE).

This change solves the issue in two separate ways:

1. Move move_emscripten_main_browser_thread_id in pthread_self_stub.c
   meaning the dependency on getpid() is more specific.
2. Avoid the use of `stderr` in `emscripten_syscall_stubs.c`