[WasmFS] Generalize test_dlmalloc_partial so it doesn't fail on extra WasmFS allocs (#16128)
The test did a printf inside a hook for each malloc. WasmFS has more allocations
(it allocates internal buffers) so it would actually infinite loop here. To avoid this
issue, which is not what the test cares about, use a lower-level logging function
and don't care about the size of the allocation.