Use stderr directly to write errors in worker.js under node. NFC
Writing directly to stderr works even when the main thread is currently
blocked (not returning to its message loop), for example in
pthread_join.  We already do this in shell.sh when defining `err` in the
primary module code.

This issue was causing error messages from the worker to be lost in the
repro case posted in #16021.