[WasmFS] Add JSImplBackend (#16209)
This refactors the existing JSFileBackend to use a new model for creating JS
backends. A JSImplBackend is defined, which adds a layer of indirection in
JS, allowing a backend to be implemented almost 100% in JS. That is,
JSImplBackend is a new backend that forwards operations to an implemenation
in JS.

See full docs in js_impl_backend.h. Note that the diffs here may not be as
useful as reading the files in their final state. In particular, look at the how the
JSFileBackend is now implemented in a minimal way in C++ in
js_file_backend.cpp.

This is a step towards it being practical to make backends like Node etc.

None of the actual JSFileBackend logic is changed, just moved around.