Don't strip all non-debug sections from separate-dwarf file
When LLVM tools such as llvm-dwp and llvm-dwarfdump parse debug info,
they use information from the wasm known sections (such as function
indices) to parse and verify the debug info. This means that #16110
broke these tools.

This is basically a revert of #16110 except that it adds tests of
dwp and dwarfdump running on separate-dwarf files. We should actually
solve #13084 again by stripping out just the code section (instead
of all known sections) but llvm-objdump currently doesn't know how
to do that (it only handles named custom sections currently).
So this PR reopens #13084