Attempt to work around Firefox CI issue. See #16242 (#16245)
I can't reproduce the issue locally, but this change got the test to pass on CI,
so I suspect the theory in #16242 is right: Firefox Dev now creates invalid
OffscreenCanvas elements in some cases. That happens on our CI, but not
locally even with the same flags, so it might be dependent on the OS or the
graphics backend or otherwise affect how Firefox creates the context. The
effect is that an OffscreenCanvas is created, but it lacks methods on the
object. To work around it, check for a valid method appearing on the object,
and if it's not there, use a normal (non-offscreen) Canvas.

Also add an assertion, which is mostly not needed assuming the workaround
is valid, but may help in the future.

Also remove -O2 in the test, which was not necessary for any reason (and it
happens to disable assertions).