Consistent use of `typeof` operator in JS library code. NFC
I noticed that in some places we were using equality checks (==/!=) and
in others we were using identity checks (===/!==).

In general identity checks are safer, but since typeof is known to
always returns a string and the RHS of these checks is also always a
string we can choose to always prefer `!=` and `==` over `!==` and `===`
for the sake of code size (sad but true, we care that much about each
  byte of output).

Perhaps a better solution would be to teach closure compiler how to
remove the extra `=` when its know that both sides of the operator are
the same type already.. but that would be a much larger change.