[WasmFS] Make Directory an abstract base class
Previously there was only a single kind of `Directory`, which kept its entries
in Wasm memory. However, other backends such as the Node or OPFS backends will
need to define their own directories. In preparation for that, make `Directory`
an abstract base class similar to `DataFile` and move its concrete in-memory
implementation into a new `MemoryDirectory` subclass similar to `MemoryFile`.

Splitting the Directory interface from its implementation required choosing what
the abstract interface should be. Many directory operations require setting or
clearing the parent pointers on added or removed child files, so to avoid each
individual backend from having to reimplement that logic, implement it on
`Directory::Handle` so that neither the underlying backends nor the clients of
the `Directory` interface have to think about it.

Since `Directory::Handle` is no longer a trivial wrapper around the underlying
`Directory` operations, move its definition out of the enclosing `Directory`
class definition to improve readability. Move the `File::Handle` and
`DataFile::Handle` definitions similarly for consistency.

Also update the interface and its users to allow directory modifications to be
fallible, which will be necessary for read-only backends.