[WasmFS] Support utime() (#16262)
This implements the __syscall_utimensat syscall and everything necessary
for that to get a basic test of utime passing.

getParsedPath() now receives an optional FD, which if given is the "base"
relative to which the path is parsed. That is what all the at* syscalls will
use.

This also switches us to use utime.c from musl instead of from JS. This is
necessary for WasmFS since the JS version only supported the JS FS. But
it works ok in the JS FS as well, as it also implements the relevant syscall,
so this is better anyhow.