Refactor mmap/munmap to store allocations in shared memory.
Previously we were storing allocation in a JS mapping which doesn't work
across threads.

With this change the implemenation of mmap/munmap is primarily in native
code, calling out of JS only for if the filesystem is needed.

This change also removes the duplicate implementation of MAP_ANON between
JS and standalone.c.

Replaces both #16297 and #16284.