Fix side module exception handling (#16309)
Exception handling for C++ code calls a bunch of functions with names like
`__cxa_find_matching_catch_##`. In Emscripten all of these functions are defined
to be the same function. Emscripten statically defines the set of these that are used
by C++ code in the main module, but it is possible that a side module will need extra
ones. This fixes the problem by updating `resolveGlobalSymbol` so that in a side module
every `__cxa_find_matching_catch_##` will resolve to `__cxa_find_matching_catch`.