[WasmFS][NFC] Rename entry to child in Directory API (#16317)
Rename `entry` to `child` whenever the API is dealing with directory contents as
shared pointers to `File`. This is in preparation for a follow-on change that
will change `getEntries` (which maintains its original name in this commit) to
return only the data needed by getdents rather than a vector of full `File`
objects.