[emscripten::val] Support literal utf-8/16 strings
It already has supports for cstrings with Latin1 encoding. And std::string
could be used for utf-8 encoded strings, but with cost on extra copy of
wire type.
I don't want to extend cstring for utf encoding so that it keeps fast in
most cases. Instead, add new static functions to specially support u8string
and u16string.