Fix Asyncify in the d8 shell (#16340)
This is similar to the Node.js path. This code is necessary
for us to throw an exception to stop execution, but keep going after it is
caught, rather than call d8's quit which halts us immediately.

Also remove some misleading annotations on tests. d8 does have setTimeout,
despite what the comments say - but perhaps it didn't in the past.