[WasmFS] Reading and writing in the Node backend (#16389)
Implement file reading and writing in the Node WasmFS backend. Since there are
no backend hooks for determining when a file is opened or closed, we cannot
store native file descriptors for open files, so we instead open and close files
each time they are read or written.

Also fix the return value of _wasmfs_node_get_mode as a drive-by.