Test more edge cases in test_unlink
Test that open files can still be read or written after they are unlinked and
that open directories can still be read after they are unlinked but that they
cannot gain new children.

Somewhat surprisingly, the old filesystem implementation passes this test with
no modifications. WasmFS cannot run it because it does not have an
implementation of `unlinkat`, and even if it did, it still would not pass.