[WasmFS] Initial pipe / poll support (#16391)
This adds the pipe and poll syscalls. Poll has more complexity but this
handles the basic cases at least.

Pipes are not that commonly used in wasm, but they do introduce the
need to mark files as seekable or not, and they are a good test of how
practical it is to make "interesting" File subtypes (which seems like it
works pretty well I think).

This enables the full test except for the fcntl part which is left for later.