[WasmFS] Track open files in the Node backend
Add callbacks to the DataFile API for notifying backends when files are opened
or closed, and call them from the OpenFileEntry constructor and destructor. In
the Node backend, open and close files on the underlying file system in these
callbacks and track the underlying file descriptor for opened files. This is
sufficient to allow the Node backend to correctly read and write opened files
that have been unlinked.

Open unlinked directories do not yet work correctly on the Node backend because
WasmFS zeroes the parent pointer on a different File object than the one held in
the open file table. Fixing that is left to a follow-on PR.