Unify `onerror` event handler used in test code. NFC
I saw this optimization/cleanup oportunity while working on #16443.

We were registering a handler both in `src/preamble.js` and in
`tests/browser_reporting.js`, but we only need one.

All the `Module['pageThrewException']` handling is in
browser_reporting.js so there is no need for preamble to be involved
here.