Fix missing `printErr` propoerty in PROXY_TO_WORKER
This issues has been causing exception on the main JS thread when
running in PROXY_TO_WORKER mode when using the default `shell.html` from
which the `printErr` override was removing back in #15555.

Since not all folks will be providing `print`/`printErr` on the
incoming module, we have to provide a fallback method.