Fix for MEMORY64 + -O3
Right now disabling `MINIFY_WASM_IMPORTS_AND_EXPORTS` requires
that `MINIFY_WASM_IMPORTED_MODULES` is also disabled.  This is
because the call to `minify_wasm_imports_and_exports` which
handles both of these processes is currently guarded by
`MINIFY_WASM_IMPORTS_AND_EXPORTS`.

In the long run we need to address this and avoid disabling
these but in the short term this fixes the broken logic.

Fixes: https://github.com/WebAssembly/binaryen/issues/4532