Never try to run proxied work without a live runtime
We were previously running `emscripten_proxy_execute_queue` from a JS callback
whenever a thread proxied to itself, even if the thread runtime had died by the
time the callback was received. In practice this never resulted in any work
running because the queue was not able to find the task queue for the thread
without the old `pthread_self` value, but it was still inconsistent with the
behavior when proxying to separate threads, which does not result in a call to
`emscripten_proxy_execute_queue`. Making the behavior more consistent lets us
insert a new assertion that `pthread_self()` is meaningful and is more robust to
future changes to the proxying mechanism.

Also improve the documentation to point out that work will be dropped if the
target thread dies with the work unfinished.